+++ changelog.md
@@ -21,6 +21,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Feature: GDPR consent popup and cookies page
 - Feature: create icon dialog now allows "tel:xxxx" and "mailto:xxxx@xxx.xxx"
 - Fix: editor icon click doesn't open link anymore
+- Fix: all urls .toLowerCase() before being saved
 
 ### [1.7.0] - 2022-10-03